¡Update 3!
BREAKING CHANGES

Includes rename of library, as the language it parses has now also
been named! =Intercept=.

Any existing build directories will no longer work, and will need to
be removed/regenerated in order for a new build to happen. This is
because CMake caches absolute paths on your local machine, and those
will have changed with everything being renamed. Best bet is to just
delete the repo and re-clone it, and follow the instructions in the
README.